+++ src/DemoFile.Game.Deadlock/Sdk/CBaseEntity.cs
@@ -1,4 +1,4 @@
-namespace DemoFile.Sdk;
+namespace DemoFile.Game.Deadlock;
 
 public partial class CBaseEntity
 {